refactored Header, Footer, Navbar
