+++ src/components/Footer/Footer.styled.js
@@ -1,4 +1,3 @@
-import React from "react"
 import styled from "styled-components"
 
 const StyledFooter = styled.footer`
@@ -32,20 +31,4 @@ const SocialMediaLink = styled.a`
   }
 `
 
-const Footer = () => (
-  <StyledFooter>
-    <SocialMediaList>
-      <SocialMediaLink href="https://github.com/filipeestacio">
-        Github
-      </SocialMediaLink>
-      <SocialMediaLink href="https://twitter.com/FilipeEstacio">
-        Twitter
-      </SocialMediaLink>
-      <SocialMediaLink href="https://www.linkedin.com/in/filipe-est%C3%A1cio-1b7bbb20/">
-        LinkedIn
-      </SocialMediaLink>
-    </SocialMediaList>
-  </StyledFooter>
-)
-
-export default Footer
+export { StyledFooter, SocialMediaList, SocialMediaLink }
