+++ src/components/Header/Header.js
@@ -1,0 +1,11 @@
+import React from "react"
+import Logo from "../logo"
+import { StyledHeader } from "./Header.styled"
+
+const Header = () => (
+  <StyledHeader>
+    <Logo />
+  </StyledHeader>
+)
+
+export default Header
