+++ src/components/Header/Header.styled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components"
+
+const StyledHeader = styled.header`
+  grid-area: header;
+  padding: 20px;
+`
+
+export { StyledHeader }
