+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import styled from "styled-components"
 import { ThemeProvider } from "styled-components"
 import { theme } from "../theme/theme"
-import Navbar from "../components/navbar"
+import Navbar from "../components/Navbar"
 import { animated, useSpring } from "react-spring"
 import { graphql } from "gatsby"
 import Helmet from "react-helmet"
